Servlets refactor
