+++ src/main/java/org/saleslist/web/PayoutServlet.java
@@ -2,47 +2,36 @@ package org.saleslist.web;
 
 import org.saleslist.model.Payout;
 import org.saleslist.repository.jdbc.JdbcPayoutRepository;
-import org.springframework.context.ConfigurableApplicationContext;
-import org.springframework.context.support.ClassPathXmlApplicationContext;
+import org.saleslist.web.user.MainServlet;
 import org.springframework.util.StringUtils;
 
 import javax.servlet.ServletConfig;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.time.temporal.ChronoUnit;
-import java.util.Objects;
 
 import static org.saleslist.web.SecurityUtil.ADMIN_ID;
 import static org.saleslist.web.SecurityUtil.getAuthUserId;
 
 @WebServlet("/payouts")
-public class PayoutServlet extends HttpServlet {
+public class PayoutServlet extends MainServlet {
 
-	private ConfigurableApplicationContext springContext;
 	private JdbcPayoutRepository payoutRepository;
 
 	@Override
 	public void init(ServletConfig config) throws ServletException {
 		super.init(config);
-		springContext = new ClassPathXmlApplicationContext("spring/spring-app.xml", "spring/spring-db.xml");
 		payoutRepository = springContext.getBean(JdbcPayoutRepository.class);
 	}
 
 	@Override
-	public void destroy() {
-		springContext.close();
-		super.destroy();
-	}
-
-	@Override
 	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-		request.setCharacterEncoding("UTF-8");
+		super.doPost(request, response);
 
 		Payout payout = new Payout(
 				LocalDateTime.parse(request.getParameter("dateTime")),
@@ -71,10 +60,8 @@ public class PayoutServlet extends HttpServlet {
 
 	@Override
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+//		super.doGet(request, response);
 		String action = request.getParameter("action");
-
-//		PayoutsStats ps = new PayoutsStats();
-//		request.setAttribute("payoutsStats", ps);
 
 		switch (action == null ? "all" : action) {
 			case "delete" -> {
@@ -98,11 +85,6 @@ public class PayoutServlet extends HttpServlet {
 				request.getRequestDispatcher("/payouts.jsp").forward(request, response);
 			}
 		}
-	}
-
-	private int getId(HttpServletRequest request) {
-		String paramId = Objects.requireNonNull(request.getParameter("id"));
-		return Integer.parseInt(paramId);
 	}
 
 	private Payout getDefaultPayout() {
